Added enough tools to allow the bot to submit random moves.
It doesn't currently check for other stones so it will crash after a while once it commits an illegal move (One that has happened already).

It also crashed because coordinate toString gives I as a value and probably an error the other way around when handed T as a column.